+++ lib/guardian_searcher/edition_result.rb
@@ -1,0 +1,41 @@
+# frozen_string_literal: true
+
+require "json"
+
+module GuardianSearcher
+  class EditionResult
+    def self.parse_with_codes(response: nil)
+      raise GuardianSearcher::GuardianSearcherUndefinedResponse unless response
+
+      case response.status
+      when 200
+        parse_results(body: response.body)
+      when 400
+        message = JSON.parse(response.body)["message"]
+        raise GuardianSearcher::GuardianBadRequestError, message
+      when 401
+        message = JSON.parse(response.body)["message"]
+        raise GuardianSearcher::GuardianUnauthorizedError, message
+      when 500
+        message = JSON.parse(response.body)["message"]
+        raise GuardianSearcher::GuardianInternalServerError, message
+      else
+        raise GuardianSearcher::GuardianUnknownError, "Unknown error, check Faraday response"
+      end
+    end
+
+    def self.parse_results(body:)
+      return unless body
+
+      body = JSON.parse(body)
+      response = body["response"]
+      GuardianSearcher::SearchResult.new(
+        current_page: response["currentPage"],
+        results: response["results"],
+        page_size: response["pageSize"],
+        pages: response["pages"],
+        start: response["startIndex"]
+      )
+    end
+  end
+end